Update Connector interfaces (#116)
This PR accomplishes a few tasks:
#99 - Document the Data Model
#114 - Connectors should accept UI State and convert to API calls
#107 - Connector shouldn't have to create ResultItem and ResultList, just particular JSON format

The purpose of it is to simplify Connectors. Connectors now simply accept the current state of the application, and convert that to an API call, and vice versa. This is much simpler than converting to and from App Search API syntax as an intermediate format as we were before.

Additionally, we dropped the usage of the custom ResultItem and ResultList types from the JavaScript api client for additional simplicity.