Fixed regereratorRuntime error
When using @babel/preset-env to target old browser vesions, it's
necessaruy to also include @babel/polyfill. I included that as a
dependency and set "useBuiltins" to "entry" which will instructs
preset to NOT include all of @babel/polyfill's polyfills, instead, only
the ones which are actually required by the target browsers.

Basically, the line "import "@babel/polyfill";" get transpiled to
individual "import xyz from "core-js" statements.

"core-js" is the underlying library that contains the polyfills.