Moved Driver init to CDM in SearchProvider
Moved SearchDriver initialization to ComponentDidMount in SearchProvider
to make sure that Server Side Rendering does not attempt to render
Search UI. Search UI is dependent on a browser environment to support
history and routing. Until we can find a better way to support
SSR, I have moved it to CDM to at least avoid errors.