Loosened the "Result" object schema
In our type definition for a "Result", we required it to have
a raw and snippet key:

```
{
  raw: "",
  snippet: ""
}
```

However, while this is the most common scenario, it's actually
technically possible to get other stuff back in a "Result".
For instance, when using the "group" feature in the App Search API,
you can get back a "_group" key in results. Or if you are using
this project with Elasticsearch, you might get a nested object back.
This schema was too arbitrarily strict. It's better to accept
anything here.