Loosened the "Result" object schema (#271)
## Description

In our type definition for a "Result", we required it to have
a raw or snippet key:

```
{
  raw: "",
  snippet: ""
}
```

However, while this is the most common scenario, it's actually
technically possible to get other stuff back in a "Result".
For instance, when using the "group" feature in the App Search API,
you can get back a "_group" key in results. Or if you are using
this project with Elasticsearch, you might get a nested object back.
This schema was too arbitrarily strict. It's better to accept
anything here.

## List of changes

- Expanded results shown in the Sandbox application, just for a better demo and also for testing.
- Updated BOTH `Result.js` types, which are duplicated in two separate packages currently, but we should definitely find a way to consolidate in the future.
- Changed the `Result.js` view to filter out arbitrary `Result` values, since we've now loosened the schema. It only attempts to render things that look like a `FieldValueWrapper` (i.e., they have a `raw` and `snippet` value, which is how App Search returns field values).

## Associated Github Issues

Fixes: #256