Fixed issue with state not updating
This fixes an issue described in #279, where if you had components
nested under a `WithSearch`, that used state from `WithSearch` to
conditionally render, they would be one state update behind.

This is because the conditional components' are initialized after
a state update on the higher level `WithSearch`, who's update was
triggered because they are subscribed to the driver's state.

When the nested components constructor runs, they grab the most
recent state from the driver. However, the way the driver was set up,
subscriptions were executed BEFORE the driver's own state was updated.
Therefore, the nested component's state was being derived from the
outdated driver state.