Fixed a CSS bug
After removing the Roboto font, there was an issue with the
ResultPerPage dropdown, where it was too narrow for the font and
was being squashed down to an ellipses.

Additionally, there was an issue where the box shadow of the
dropdown was being cut off due to a change made in #266. I moved
the styles from that change to be more targeted to fix this.