Updated setSearchTerm tests
I updated these tests because they were testing the wrong thing.

There are two side effects that occur as a result of setSearchTerm:
_udpateSearchResults
_makeSearchRequest

_udpateSearchResults is what we are debouncing, so we want to verify
whether that has run or not. Testing whether or not an API request
has been make is not a good way to test this, since _makeSearchRequest
is now always debounced.

In other words, we could call _udpateSearchResults 3 times in a row
and it will execute 3 times, but ultimately only one api call will be
made.