Bugfix: Facets break on an empty string
Facets were breaking both in the UI with a javascript error, and
on the server with a 4xx when documents contained fields with empty
string values.

I believe the same thing would happen for any falsy value in a document
field, including `0`.

In the App Search responseAdapter, when attempting to build our facet
objects in state, it was tripping a falsy check which made the code
perform logic as if the returned value was an object (like a range
facet) rather than a raw value.

For example, the following facet returned by the server:

```
{
  type: "value",
  data: [
    {
      value: "",
      count: 5
    }
  ]
}
```

should be converted to:

```
{
  type: "value",
  field: "states",
  data: [
    {
      value: "",
      count: 5
    }
  ]
}
```

But instead was converted to:
```
{
  type: "value",
  field: "states",
  data: [
    {
      value: {}, // This is incorrect
      count: 5
    }
  ]
}
```

After fixing the above mentioned issue, I realized that there were
problems in MultiCheckboxFacet.

The MultiCheckboxFacet's logic for determining which value is selected
was doing a truthy-based conversion to a boolean value when finding
values, which converted `""` to `false`. The `false` flag indicated
that that particular facet value was not the selected facet value, even
though it was.

For example:

```
selectedValue = ""

[ ] "Arizona"
[ ] "" <-- Should be marked as selected
[ ] "Utah"
```

After fixing that, I also realized that we were using a `deepEqual`
check, which used `==` under the hood to determine equality. That means
that if a value like `0` was the selected value, that it would
incorrectly match `==`.

For example:

```
selectedValue = ""

[x] 0 <-- Should not be marked as selected
[ ] 1
[ ] 2
```