Refactor filter matching logic
In order to properly write enough tests to feel confident about
this logic, I had to factor out filter value matching logic
into a helper.

This logic was duplicated in both search-ui and
react-search-ui-views. I exported a helper for this from search-ui to
react-search-ui, and passed it in as a prop into the Facet
views.

Passing helpers into views makes common methods
readily available to the view without requiring them to import
them from anywhere. It makes creating custom views easier, as
common behaviors are abstracted out to functions that you can
easily use to compose the view.