fix: Replace debounce-fn in DebounceManager to facilitate IE11 compat… (#442)
Even after adding polyfills in support of IE11, search-ui is still unable to render properly due to an issue in using Object.defineProperty() when mimic-fn a function as part of the use of debounce-fn.

On our own fork of search-ui, we've replaced the use of the debounce-fn module with our own debounce function. After discussion with my team, we wanted to submit a PR with our debounce function so that we wouldn't have to maintain our own fork specifically for IE11 compatibility.