fix: Search-UI smart url forces extra history push
Currently, when showing search results on page load, an extra history
event is added to history, causing a few issues, including the
inconvenience on users of having to press the back button twice to get
back to where they were originally.

The fix is to simply use a browser history "replace" instead of "push"
on the any queries happening on page load.

This PR is based on the following PR:
https://github.com/elastic/search-ui/pull/443

Fixes https://github.com/elastic/search-ui/issues/434