Added ResultWithRequiredId type
Some views actually required an "id" property on result objects
to function properly, which was not expressed in validations.

I couldn't simply turn on an "id" requirement for Result across
the board because it could break certain views that actually
don't require it.

For that reason, I create a separate view, ResultWithRequiredId, which
enforces the "id" requirement.

I also added a special "TODO Next Major" to help indicate breaking
changes that should be made in the next major.