Re-arranged dev deps for Hooks
In order to use hooks within the local sandbox environment,
we need to resolve an "invalid hook call" error.

A number of solutions are proposed here:
https://reactjs.org/warnings/invalid-hook-call-warning.html

In our case, the error stemmed from the fact that we link our
dependencies from sandbox to our packages. To resolve that, we
need to also link the `react` in our libraries to the version
that sandbox uses.

I chose to add that linking automatically so that it happens
when running `npm run sandbox`, to make it seamless for developers.