This commit attempts to simplify views
When users are creating custom views, they are instructed
to copy and paste the existing view from source and customize
it.

currently, there are a number of external imports that make the
copy/paste experience difficult.

This commit changes these views to all accept a "viewHelpers" prop,
which contain all view helpers for use. This means they can be
easily used without figuring out how to deal with imports.

This also had the side benefit for allowing us to delete duplicated
types throughout the two projects. I simply changed to the prop type
checks to be simpler in the view project and just use primitive
prop types.