Limited App Search paging to 100 results
There is a known issue in the App Search API in which a search response
may return a total page count of over 100, even thought the API can
only page up to page 100.

This will break Search UI's paging, which assumes that if the API
says there are 1000 pages, that they can all be paged through.

This commit rewrites the response in the App Search Connector to cap
pages to 100.