Update tutorials-elasticsearch.asciidoc
Typo line 483,  facet keys need to be unique there are 2 number 4's. Changed the last one to a 5.

resulting in error as such otherwise 

```react-dom.development.js:67 Warning: Encountered two children with the same key, `4`. Keys should be unique so that components maintain their identity across updates. Non-unique keys may cause children to be duplicated and/or omitted — the behavior is unsupported and could change in a future version.
at div
...< truncated remaining error msg>